Add main() guards to CLI entry-point scripts #220
